requirements.txt
asgiref==3.11.1
boto3==1.42.46
botocore==1.42.46
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
dj-database-url==3.1.0
Django==6.0.2
django-admin-honeypot-updated-2021==1.2.0
django-cleanup==9.0.0
django-imagekit==6.0.0
django-environ==0.12.0
django-htmx==1.27.0
django-ipware==7.0.1
django-storages==1.14.6
gunicorn==25.0.3
idna==3.11
jmespath==1.1.0
packaging==26.0
pillow==12.1.1
psycopg2-binary==2.9.11
pycparser==3.0
python-dateutil==2.9.0.post0
python-ipware==3.0.0
requests==2.32.5
s3transfer==0.16.0
six==1.17.0
sqlparse==0.5.5
urllib3==2.6.3
whitenoise==6.11.0
